refactor: extract optimistic sessions from AgentWorkspace
